test: Split post tests and write post patch default color test.
